% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{cat_numeric}
\alias{cat_numeric}
\title{cat_numeric: Categorization of numeric vector with logical defaults and formatting}
\usage{
cat_numeric(x, cutoffs = c(6, 18, 41, 66), start = -1, stop = 1000)
}
\arguments{
\item{x}{Numeric vector}

\item{cutoffs}{Numeric vector with cutoffs}

\item{start}{Lower value to include in first level}

\item{stop}{Upper value to include in last level}
}
\value{
A factor with categorized values
}
\description{
cat_numeric: Categorization of numeric vector with logical defaults and formatting
}
\examples{
cat_numeric(1:10, cutoffs = c(2.5, 50, 7.5, 10), stop = 11)

}
