% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_mean.R
\name{ci_mean}
\alias{ci_mean}
\title{Get confidence intervals for mean}
\usage{
ci_mean(x, level = 0.95)
}
\arguments{
\item{x}{A numeric vector}

\item{level}{Confidence level, Default: 0.95}
}
\value{
A vector with c(cil, ciu)
}
\description{
Pass a numeric vector and get vector with confidence interval for the mean
}
\details{
Uses lm and confint to get interval
}
\examples{
ci_mean(1:10)
}
