% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_output.R
\name{colorize_span}
\alias{colorize_span}
\title{colorize_span}
\usage{
colorize_span(text, color = "red")
}
\arguments{
\item{text}{Text}

\item{color}{Color of the output. Hex or named.}
}
\value{
Returns an span tag with a color or Latex colored text
}
\description{
colorize_span
}
