% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{convert_idate_to_date}
\alias{convert_idate_to_date}
\title{convert_idate_to_date to lubridate::ymd in a data.frame}
\usage{
convert_idate_to_date(data)
}
\arguments{
\item{data}{A data.frame}
}
\value{
A data.frame with all iDates converted
}
\description{
convert_idate_to_date to lubridate::ymd in a data.frame
}
