% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{convert_sas_dates_to_date}
\alias{convert_sas_dates_to_date}
\title{convert_sas_dates_to_date: Convert all sas dates in a data.frame to lubridate::ymd}
\usage{
convert_sas_dates_to_date(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A data.frame with all sas dates converted
}
\description{
convert_sas_dates_to_date: Convert all sas dates in a data.frame to lubridate::ymd
}
