% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_tableone.R
\name{convert_tableone}
\alias{convert_tableone}
\title{Convert TableOne Object to Data Frame}
\usage{
convert_tableone(tabone, ...)
}
\arguments{
\item{tabone}{A \code{TableOne} object to be converted.}

\item{...}{Additional arguments passed to the \code{print} function. See ?print.TableOne for options}
}
\value{
A data frame representation of the \code{TableOne} object.
}
\description{
This function converts a \code{TableOne} object into a data frame.
}
\examples{
\dontrun{
tabone <- CreateTableOne(vars = c("age", "sex"), data = mydata)
df <- convert_tableone(tabone)
}
}
