% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_package_utils.R
\name{create_r_and_test}
\alias{create_r_and_test}
\title{create_r_and_test}
\usage{
create_r_and_test(file_name, should_open = FALSE)
}
\arguments{
\item{file_name}{Name of the file to crate}

\item{should_open}{Logical, should files be opened?}
}
\value{
None
}
\description{
create_r_and_test
}
