% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated_all.R
\name{duplicated_all}
\alias{duplicated_all}
\title{duplicated_all}
\usage{
duplicated_all(x)
}
\arguments{
\item{x}{A vector of elements or a data.table}
}
\value{
A logical vector indicating ALL duplicated elements in x or rows if data.table
}
\description{
duplicated_all
}
\examples{
duplicated_all(c(rep("a", 3), "b", "c"))
}
