% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated_all.R
\name{duplicated_values}
\alias{duplicated_values}
\title{duplicated values}
\usage{
duplicated_values(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A vector of duplicated items or a data.table with duplicated rows
}
\description{
duplicated values
}
\examples{
duplicated_values(c(rep("a", 3), "b", "c"))

}
