% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_case_when.r
\name{fct_case_when}
\alias{fct_case_when}
\title{Factor case when}
\usage{
fct_case_when(...)
}
\arguments{
\item{...}{Arguments passed to \code{dplyr::case_when()}}
}
\value{
A factor with the same order of levels as the cases in case when
}
\description{
Use case when to create a factor
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[stringr]{str_remove}}
\code{\link[dplyr]{case_when}}
}
