% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patreg.R
\name{find_icd_codes_from_patreg}
\alias{find_icd_codes_from_patreg}
\title{find_icd_codes_from_patreg}
\usage{
find_icd_codes_from_patreg(patreg, icds_regex)
}
\arguments{
\item{patreg}{Patreg data.frame from SoS}

\item{icds_regex}{A regex with icd codes to match on}
}
\value{
Data frame from patreg matching any of the kva codes
}
\description{
find_icd_codes_from_patreg
}
