% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patreg.R
\name{find_kva_codes_from_patreg}
\alias{find_kva_codes_from_patreg}
\title{find_kva_codes_from_patreg}
\usage{
find_kva_codes_from_patreg(patreg, kva_regex)
}
\arguments{
\item{patreg}{Patreg data.frame from SoS}

\item{kva_regex}{A regex with kva codes to match on}
}
\value{
Data frame from patreg matching any of the kva codes
}
\description{
find_kva_codes_from_patreg
}
