% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_pattern_in_columns.R
\name{find_pattern_in_columns}
\alias{find_pattern_in_columns}
\title{find_pattern_in_columns}
\usage{
find_pattern_in_columns(df, pattern)
}
\arguments{
\item{df}{A data.frame object}

\item{pattern}{Pattern to find in columns}
}
\value{
A logical vector indicating if pattern was found in row
}
\description{
find_pattern_in_columns
}
\examples{
find_pattern_in_columns(iris[1:10, ], "setosa")

}
