% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_variable_in_table.R
\name{find_variable_in_table}
\alias{find_variable_in_table}
\title{find_variable in table}
\usage{
find_variable_in_table(df, var_pattern, ignore_case = TRUE)
}
\arguments{
\item{df}{A dataframe to find the variable in}

\item{var_pattern}{The regex pattern to match variable}

\item{ignore_case}{Should case be ignored?}
}
\value{
A character vector with all matching column names
}
\description{
find_variable in table
}
\examples{
find_variable_in_table(iris, "species")

}
