% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_packages.R
\name{install_packages}
\alias{install_packages}
\title{Install Common R Packages}
\usage{
install_packages(packages = NULL, quiet = FALSE)
}
\arguments{
\item{packages}{Character vector of additional package names to check/install}

\item{quiet}{Logical indicating whether to suppress messages (default: FALSE)}
}
\value{
Invisible character vector of newly installed packages
}
\description{
This function checks if specified packages are installed and installs
missing packages from CRAN. It comes with a default set of commonly used packages
and any additional packages specified will be added to these defaults.
}
