% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{is_na}
\alias{is_na}
\title{is_na}
\usage{
is_na(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A logical vector indicating if elements are NA or ""
}
\description{
is_na
}
\examples{
is_na(c(rep(NA, 2), rep("", 2), "x"))
}
