% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_merge.R
\name{left_merge}
\alias{left_merge}
\title{left_merge}
\usage{
left_merge(x, y, by = NULL, id_vars = "lopnr")
}
\arguments{
\item{x}{A data.frame lhs}

\item{y}{A data.frame rhs}

\item{by}{A character vector to join on}

\item{id_vars}{A character vector to join on. "lopnr" by default}
}
\value{
A left joined table where duplicated columns are removed.
}
\description{
left_merge
}
