% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{mean_sd}
\alias{mean_sd}
\title{mean_sd}
\usage{
mean_sd(x, remove_na = TRUE, digits = 1)
}
\arguments{
\item{x}{Numeric vector}

\item{remove_na}{Should NAs be removed?}

\item{digits}{The number of decimal places to include}
}
\value{
A string of form mean (sd)
}
\description{
mean_sd
}
\examples{
mean_sd(1:10)

}
