% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{median_iqr}
\alias{median_iqr}
\title{median_iqr}
\usage{
median_iqr(x, remove_na = TRUE, digits = 1)
}
\arguments{
\item{x}{Numeric vector}

\item{remove_na}{Should NAs be removed?}

\item{digits}{The number of decimal places to include}
}
\value{
A string of form median (iqr)
}
\description{
median_iqr
}
\examples{
median_iqr(1:10)

}
