% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{n_per}
\alias{n_per}
\title{n_per}
\usage{
n_per(x, remove_na = FALSE, ...)
}
\arguments{
\item{x}{Numeric or logical vector of form 1/0 or TRUE/FALSE}

\item{remove_na}{Should NAs be removed?}

\item{...}{Additional arguments passed to scales::percent}
}
\value{
A string of form count (\%)
}
\description{
n_per
}
\examples{
n_per(c(rep(1, 3), rep(0, 3)))

}
