% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{n_prop}
\alias{n_prop}
\title{n_prop}
\usage{
n_prop(n, denom, accuracy = 1)
}
\arguments{
\item{n}{Nominator}

\item{denom}{Denominator}

\item{accuracy}{Which accuracy should be used}
}
\value{
A character of form n (perc\%)
}
\description{
n_prop
}
\examples{
n_prop(1, 10)
}
