% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{pcollapse}
\alias{pcollapse}
\title{pcollapse}
\usage{
pcollapse(x, collapse = " | ")
}
\arguments{
\item{x}{A vector of elements to be combined}

\item{collapse}{Seperator used for collapsing the elements}
}
\value{
A quoted combination of x seperated by sep
}
\description{
pcollapse
}
\examples{
pcollapse(letters[1:5])
}
