% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{perc}
\alias{perc}
\title{perc}
\usage{
perc(x, accuracy = 0.1, ...)
}
\arguments{
\item{x}{A proportion}

\item{accuracy}{Use 0.1 for 1 decimal, 1 for 0 decimals, 0.01 for 2 etc.}

\item{...}{Additional arguments passed to percent}
}
\value{
A character of form "10\%"
}
\description{
perc
}
\examples{
perc(0.1)
}
