% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{roundn}
\alias{roundn}
\title{roundn}
\usage{
roundn(x, dig = 2)
}
\arguments{
\item{x}{A float value to round and format}

\item{dig}{Number of digits to round to}
}
\value{
A rounded value of as string with the correct number of digits
}
\description{
roundn
}
\examples{
roundn(1.100100010, dig = 3)
}
