% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{simplify_colnames}
\alias{simplify_colnames}
\title{simplify_colnames: Helper to remove non-ascii and weird characters}
\usage{
simplify_colnames(data)
}
\arguments{
\item{data}{Data.frame which colnames to change}
}
\value{
A data.frame with simplified colnames
}
\description{
simplify_colnames: Helper to remove non-ascii and weird characters
}
\examples{
simplify_colnames(iris)
}
