% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{split_collapse}
\alias{split_collapse}
\title{split_collapse}
\usage{
split_collapse(x, sep = " | ")
}
\arguments{
\item{x}{A character of collapsed items}

\item{sep}{Seperator used for seperations}
}
\value{
A seperated character vector
}
\description{
split_collapse
}
\examples{
split_collapse("a | b")
}
