% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_output.R
\name{table_dt}
\alias{table_dt}
\title{table_dt: Create nice looking datatable with sane defaults}
\usage{
table_dt(
  data,
  col_names = NULL,
  first_colname = NULL,
  title_row_names = TRUE,
  title_col_names = TRUE,
  row_groups = FALSE,
  alignment = "dt-center",
  align_targets = NULL,
  page_length = 20,
  class = NULL,
  row_callback = NULL,
  width = NULL,
  ...
)
}
\arguments{
\item{data}{Table to present}

\item{col_names}{New column names to be used in table}

\item{first_colname}{A character vector with new colnames}

\item{title_row_names}{Should first row be formatted?}

\item{title_col_names}{Should colnames be formatted?}

\item{row_groups}{Should first columns be grouped?}

\item{alignment}{Alignment of cells c("dt-center", "dt-right", "dt-left")}

\item{align_targets}{Numeric vector containing columns positions to}

\item{page_length}{Number of rows per page}

\item{class}{Classes to apply. Defaults to "compact stripe"}

\item{row_callback}{Row callback to apply.}

\item{width}{Width as proportion of parent div. Based on ncol bby default}

\item{...}{Additional variables to pass to datatable::DT}
}
\value{
An html table
}
\description{
table_dt: Create nice looking datatable with sane defaults
}
