% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose_aggregate_dt.R
\name{trans_dt}
\alias{trans_dt}
\title{Transpose as wide dt}
\usage{
trans_dt(x)
}
\arguments{
\item{x}{A table with variables in columns and groups as rows}
}
\value{
A table with variables as rows and groups as columns
}
\description{
Format it like a proper long table
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[dplyr]{pull}}, \code{\link[dplyr]{bind_cols}}
\code{\link[tibble]{tibble}}
}
