% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_manipulation.R
\name{transform_aggregate}
\alias{transform_aggregate}
\title{transpose_aggregate Transpose the way I want}
\usage{
transform_aggregate(.data)
}
\arguments{
\item{.data}{Table to be fransformed}
}
\value{
A transposed table
}
\description{
transpose_aggregate Transpose the way I want
}
