% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose_aggregate_dt.R
\name{transpose_aggregated_dt}
\alias{transpose_aggregated_dt}
\title{transpose_aggregate Transpose the way I want}
\usage{
transpose_aggregated_dt(table1)
}
\arguments{
\item{table1}{Table to be fransformed}
}
\value{
A transposed table
}
\description{
transpose_aggregate Transpose the way I want
}
