% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{truethy}
\alias{truethy}
\title{truethy}
\usage{
truethy(x, trueth = "Ja")
}
\arguments{
\item{x}{A logical, character or numeric vector. If character, use trueth to indicate TRUE value #nolint}

\item{trueth}{Optional, a string indicating the TRUE value.}
}
\value{
A logical vector, treating NA ans "" and 0 as FALSE.
}
\description{
truethy
}
\examples{
truethy(c(TRUE, TRUE, FALSE, NA))
}
