% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{x_ci}
\alias{x_ci}
\title{x_ci}
\usage{
x_ci(x, cil, ciu, dig = 1)
}
\arguments{
\item{x}{A centric measure}

\item{cil}{Lower confidence value}

\item{ciu}{Upper confidence value}

\item{dig}{Number of digits to round to}
}
\value{
A string for form x (cil-ciu)
}
\description{
x_ci
}
\examples{
x_ci(1.0101, 0.500, 1.353, dig = 2)

}
