% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{year_month}
\alias{year_month}
\title{year_month}
\usage{
year_month(date)
}
\arguments{
\item{date}{A date}
}
\value{
A string of form YYYY-MM
}
\description{
year_month
}
\examples{
year_month(as.Date("2020-01-01"))

}
